+++ src/gui/CraftingPage.lua
@@ -9,9 +9,9 @@ local Self = GUI.CraftingPage
 
 ---@enum CraftingPage.Filter
 Self.Filter = {
-    Sorted = "SORTED",
+    Scan = "SCAN",
+    Queue = "QUEUE",
     Restock = "RESTOCK",
-    Tracked = "TRACKED",
 }
 
 ---@type CraftingPage.Filter?
@@ -27,10 +27,11 @@ end)
 ---@type TreeDataProviderMixin
 Self.dataProvider = CreateTreeDataProvider()
 
-Self.dataProvider:SetSortComparator(function(a, b)
+Self.dataProvider:SetSortComparator(function(a, b) --[[@cast a RecipeTreeNode]] --[[@cast b RecipeTreeNode]]
     if not a or not b then return a ~= b and b == nil end
     a, b = a:GetData(), b:GetData()
     if (a.amount == 0) ~= (b.amount == 0) then return a.amount ~= 0 end
+    if not a.value or not b.value then return a.recipeInfo.name < b.recipeInfo.name end
     if Self:GetFilterSort(Self.filter) == Optimization.Method.Cost then return a.value < b.value end
     return a.value > b.value
 end, false, true)
@@ -62,28 +63,34 @@ end
 ---------------------------------------
 
 ---@param frame ProfessionsRecipeListRecipeFrame
----@param node TreeNodeMixin
+---@param node RecipeTreeNode
 function Self:InitRecipeListRecipe(frame, node)
     if not frame.GetLabelColor then return end
 
-    local data = node:GetData() --[[@as RecipeTreeNodeData]]
+    local data = node:GetData()
     local value, amount, quality = data.value, data.amount, data.quality
     local r, g, b = frame:GetLabelColor():GetRGB()
 
-    frame:HookScript("OnClick", function (_, buttonName)
-        if buttonName ~= "LeftButton" or not IsModifiedClick("RECIPEWATCHTOGGLE") then return end
+    -- Hook OnClick
+    if not frame.hooked then
+        frame.hooked = true
 
-        if self.filter == self.Filter.Restock then
-            local operation, quality, amount = data.operation, data.quality, data.amount
-            local recipe = operation.recipe
+        frame:HookScript("OnClick", function (_, buttonName)
+            if buttonName ~= "LeftButton" or not IsModifiedClick("RECIPEWATCHTOGGLE") then return end
 
-            if amount <= 0 or not Recipes:IsTracked(recipe) then return end
+            if self.filter == self.Filter.Restock then
+                local operation, quality, amount = data.operation, data.quality, data.amount
+                if not operation or not amount then return end
+                local recipe = operation.recipe
 
-            Recipes:SetTrackedPerQuality(recipe, quality ~= nil)
-            Recipes:SetTrackedAmount(recipe, amount, quality)
-            Recipes:SetTrackedAllocation(recipe, operation, quality)
-        end
-    end)
+                if amount <= 0 or not Recipes:IsTracked(recipe) then return end
+
+                Recipes:SetTrackedPerQuality(recipe, quality ~= nil)
+                Recipes:SetTrackedAmount(recipe, amount, quality)
+                Recipes:SetTrackedAllocation(recipe, operation, quality)
+            end
+        end)
+    end
 
     -- Set amount
     if amount then
@@ -170,8 +177,8 @@ end
 function Self:ModifyRecipeListFilter()
     local IsFilterSelected = function (filter) return self:IsFilterSelected(filter) end
     local SetFilterSelected = function (filter) self:SetFilterSelected(filter) return 4 end
-    local IsSortSelected = function (sort) return self:IsFilterSelected(self.Filter.Sorted, sort) end
-    local SetSortSelected = function (sort) self:SetFilterSelected(self.Filter.Sorted, sort) return 4 end
+    local IsSortSelected = function (sort) return self:IsFilterSelected(self.Filter.Scan, sort) end
+    local SetSortSelected = function (sort) self:SetFilterSelected(self.Filter.Scan, sort) return 4 end
 
     Menu.ModifyMenu("MENU_PROFESSIONS_FILTER", function (_, rootDescription)
         if ProfessionsFrame:GetTab() ~= ProfessionsFrame.recipesTabID then return end
@@ -181,20 +188,22 @@ function Self:ModifyRecipeListFilter()
         rootDescription:Insert(MenuUtil.CreateRadio(NONE, IsFilterSelected, SetFilterSelected))
 
         -- Add sort menu
-        local sortSubmenu = rootDescription:Insert(MenuUtil.CreateRadio("Scan", IsFilterSelected, SetFilterSelected, self.Filter.Sorted)) --[[@as ElementMenuDescriptionProxy]]
+        if Prices:IsSourceInstalled() then
+            local sortSubmenu = rootDescription:Insert(MenuUtil.CreateRadio("Scan", IsFilterSelected, SetFilterSelected, self.Filter.Scan)) --[[@as ElementMenuDescriptionProxy]]
 
-        for name,method in pairs(Optimization.Method) do repeat
-            if method == Optimization.Method.CostPerConcentration then
-                break
-            elseif method == Optimization.Method.ProfitPerConcentration then
-                name = "Profit per Concentration"
-            end
+            for name,method in pairs(Optimization.Method) do repeat
+                if method == Optimization.Method.CostPerConcentration then
+                    break
+                elseif method == Optimization.Method.ProfitPerConcentration then
+                    name = "Profit per Concentration"
+                end
 
-            sortSubmenu:CreateRadio(name, IsSortSelected, SetSortSelected, method)
-        until true end
+                sortSubmenu:CreateRadio(name, IsSortSelected, SetSortSelected, method)
+            until true end
+        end
 
         -- Add tracked and restock options
-        rootDescription:Insert(MenuUtil.CreateRadio("Queue", IsFilterSelected, SetFilterSelected, self.Filter.Tracked))
+        rootDescription:Insert(MenuUtil.CreateRadio("Queue", IsFilterSelected, SetFilterSelected, self.Filter.Queue))
         rootDescription:Insert(MenuUtil.CreateRadio("Restock", IsFilterSelected, SetFilterSelected, self.Filter.Restock))
     end)
 
@@ -213,21 +222,23 @@ end
 ---@param filter? CraftingPage.Filter
 ---@param sort? Optimization.Method
 function Self:SetFilterSelected(filter, sort)
-    if filter == self.Filter.Sorted then
+    if filter == self.Filter.Scan then
         sort = sort or self.sort or Optimization.Method.Profit
         GUI.RecipeForm.CraftingForm:SetOptimizationMethod(sort)
     else
         sort = filter and self.sort or nil
     end
 
+    if filter == self.filter and sort == self.sort then return end
+
     self.filter = filter
     self.sort = sort
 
-    self.professionInfo, self.recipeIDs, self.selectedRecipeID, self.selectedQuality = nil, nil, nil, nil
+    self.dataProvider:GetRootNode():Flush()
 
     if filter then ---@cast sort -?
         self.frame.RecipeList:SetPoint("BOTTOMLEFT", 0, 32.5)
-        self:UpdateRecipeList(true)
+        self:UpdateRecipeList(true, true)
     else
         self.frame.RecipeList:SetPoint("BOTTOMLEFT", 0, 5)
         self:UpdateFilterButtons()
@@ -241,64 +252,106 @@ end
 
 ---@param filter? CraftingPage.Filter
 function Self:GetFilterSort(filter)
-    return filter == self.Filter.Sorted and self.sort or filter and Optimization.Method.Profit or nil
+    return filter == self.Filter.Scan and self.sort or filter and Optimization.Method.Profit or nil
 end
 
 ---@param refresh? boolean
-function Self:UpdateRecipeList(refresh)
+---@param flush? boolean
+function Self:UpdateRecipeList(refresh, flush)
     if not self.filter then return end
     local filter, sort = self.filter, self:GetFilterSort(self.filter) --[[@cast filter -?]] --[[@cast sort -?]]
+    local applyConcentration = Util:OneOf(sort, Optimization.Method.CostPerConcentration, Optimization.Method.ProfitPerConcentration)
 
     local profInfo = self.frame.professionInfo
     local professionChanged = not self.professionInfo or not Util:TblMatch(self.professionInfo, "professionID", profInfo.professionID, "skillLevel", profInfo.skillLevel, "skillModifier", profInfo.skillModifier)
 
     local recipeIDs = self:GetFilterRecipeIDs(filter)
     local n = #recipeIDs
-    local filterChanged = not self.recipeIDs or not Util:TblEquals(self.recipeIDs, recipeIDs)
 
     self.professionInfo = profInfo
     self.recipeIDs = recipeIDs
 
     local recipeList = self.frame.RecipeList
+    local scrollBox = recipeList.ScrollBox
     local noResultsText = recipeList.NoResultsText
     local selectionBehavior = recipeList.selectionBehavior
     local provider = self.dataProvider
+    local hasCustomProvider = scrollBox:GetDataProvider() == provider
 
-    if professionChanged or recipeList.ScrollBox:GetDataProvider() == provider then
-        local selected = selectionBehavior:GetFirstSelectedElementData() --[[@as TreeNodeMixin]]
+    local updateSelected = professionChanged or hasCustomProvider
+    local flush = flush or professionChanged
+    local refresh = refresh
+        or professionChanged
+        or not hasCustomProvider
+        or not self.recipeIDs or not Util:TblEquals(self.recipeIDs, recipeIDs)
+
+    -- Update selected
+    if updateSelected then
+        local selected = selectionBehavior:GetFirstSelectedElementData() --[[@as RecipeTreeNode]]
         if selected then
-            local data = selected:GetData() --[[@as RecipeTreeNodeData]]
+            local data = selected:GetData()
             self.selectedRecipeID, self.selectedQuality = data.recipeInfo.recipeID, data.quality
         else
             self.selectedRecipeID, self.selectedQuality = recipeList.previousRecipeID, nil
         end
     end
 
-    if refresh or professionChanged or filterChanged then
+    -- Flush or clear data provider
+    if flush then
         provider:GetRootNode():Flush()
-        recipeList.ScrollBox:SetDataProvider(provider, ScrollBoxConstants.RetainScrollPosition)
-        selectionBehavior:ClearSelections()
+    else
+        local recipeIDMap = Util:TblFlip(recipeIDs)
+        for _,node in provider:EnumerateEntireRange() do repeat ---@cast node RecipeTreeNode
+            local data = node:GetData()
+            if recipeIDMap[data.recipeInfo.recipeID] then break end
+            provider:Remove(node)
+        until true end
+    end
 
+    -- Set data provider
+    if not hasCustomProvider then
+        scrollBox:SetDataProvider(provider, ScrollBoxConstants.RetainScrollPosition)
+        selectionBehavior:ClearSelections()
+    end
+
+    -- Refresh recipes
+    if refresh then
         if professionChanged then self.filterCache:Clear() end
 
         Promise:Async(function ()
             Promise:GetCurrent():SetPriority(5)
 
-            for i,recipeID in ipairs(recipeIDs) do
-                local recipe = C_TradeSkillUI.GetRecipeSchematic(recipeID, false)
+            ---@type table<TreeNodeMixin, true>
+            local datas = {}
+
+            for i,recipeID in ipairs(recipeIDs) do repeat
+                wipe(datas)
+
                 local recipeInfo = C_TradeSkillUI.GetRecipeInfo(recipeID)
+                if not recipeInfo then break end
+
+                local recipe = C_TradeSkillUI.GetRecipeSchematic(recipeID, recipeInfo.isRecraft)
                 local operations = self:GetFilterRecipeOperations(filter, recipe, sort)
 
                 if operations then
                     for _,operation in pairs(operations) do repeat
-                        local quality = recipeInfo and recipeInfo.supportsQualities and operation:GetResultQuality() or nil
-                        local value = Optimization:GetOperationValue(operation, sort)
+                        local quality = recipeInfo.supportsQualities and operation:GetResultQuality() or nil
 
-                        if not value or abs(value) == math.huge then break end
+                        local value
+                        if Prices:IsSourceInstalled() then
+                            value = Optimization:GetOperationValue(operation, sort)
+                            if not value or abs(value) == math.huge then break end
+                        end
 
                         local amount, amountShown, amountTotal = self:GetFilterRecipeAmount(filter, recipe, quality, value)
 
-                        provider:Insert({
+                        local node = provider:FindElementDataByPredicate(function (node) ---@cast node RecipeTreeNode
+                            local data = node:GetData()
+                            return data.recipeInfo.recipeID == recipeID and data.quality == quality
+                        end, false) --[[@as RecipeTreeNode?]]
+
+                        ---@type RecipeTreeNodeData
+                        local data = node and node:GetData() or {
                             recipeInfo = recipeInfo,
                             operation = operation,
                             value = value,
@@ -307,18 +360,43 @@ function Self:UpdateRecipeList(refresh)
                             amount = amount,
                             amountShown = amountShown,
                             amountTotal = amountTotal
-                        })
+                        }
+
+                        -- Insert or update node
+                        if not node then
+                            provider:Insert(data)
+                        else
+                            data.operation = operation
+                            data.value = value
+                            data.method = sort
+                            data.quality =  quality
+                            data.amount = amount
+                            data.amountShown = amountShown
+                            data.amountTotal = amountTotal
+
+                            local frame = scrollBox:FindFrame(node) --[[@as ProfessionsRecipeListRecipeFrame]]
+                            if frame then self:InitRecipeListRecipe(frame, node) end
+                        end
+
+                        datas[data] = true
                     until true end
                 end
 
-                if recipeID == self.selectedRecipeID and not selectionBehavior:HasSelection() then
-                    Promise:Yield(i, n, recipeID) -- Yield so UI can be updated
-                else
-                    Promise:YieldProgress(i, n, recipeID)
-                end
-            end
+                -- Remove unused qualities
+                for _,node in provider:EnumerateEntireRange() do repeat ---@cast node RecipeTreeNode
+                    local data = node:GetData()
+                    if data.recipeInfo.recipeID ~= recipeID or datas[data] then break end
+                    provider:Remove(node)
+                until true end
 
-            -- Fix last recipe not getting sorted correctly
+                -- Update progress and maybe yield
+                if not hasCustomProvider and recipeID == self.selectedRecipeID and not selectionBehavior:HasSelection() then
+                    Promise:Yield(i, n, recipeID) -- Yield to restore selection
+                else
+                    Promise:YieldProgress(i, n)
+                end
+            until true end
+
             provider:Invalidate()
         end):Singleton(self, "filterJob"):Start(function ()
             noResultsText:SetShown(false)
@@ -333,17 +411,16 @@ function Self:UpdateRecipeList(refresh)
 
             self.progressBar:Progress(i, n)
 
-            if recipeID == self.selectedRecipeID and not selectionBehavior:HasSelection() then
-                self:RestoreSelectedRecipe()
-            end
+            if not recipeID then return end
+
+            self:RestoreSelectedRecipe()
         end):Finally(function ()
             noResultsText:SetShown(provider:IsEmpty())
 
             self:UpdateFilterButtons()
         end)
     else
-        recipeList.ScrollBox:SetDataProvider(provider, ScrollBoxConstants.RetainScrollPosition)
-        self:RestoreSelectedRecipe()
+        provider:Invalidate()
     end
 end
 
@@ -353,8 +430,8 @@ function Self:RestoreSelectedRecipe()
     local recipeID, quality = self.selectedRecipeID, self.selectedQuality
     if not recipeID then return end
 
-    self.frame.RecipeList.selectionBehavior:SelectFirstElementData(function (node) ---@cast node TreeNodeMixin
-        local data = node:GetData() --[[@as RecipeTreeNodeData]]
+    self.frame.RecipeList.selectionBehavior:SelectFirstElementData(function (node) ---@cast node RecipeTreeNode
+        local data = node:GetData()
         return data.recipeInfo.recipeID == recipeID and (not data.quality or not quality or data.quality == quality)
     end)
 end
@@ -363,13 +440,13 @@ end
 function Self:GetFilterRecipeIDs(filter)
     local recipeIDs = C_TradeSkillUI.GetFilteredRecipeIDs() ---@cast recipeIDs -?
 
-    if filter == self.Filter.Sorted then
+    if filter == self.Filter.Scan then
         local professionID = self.frame.professionInfo.professionID
 
         return Util:TblFilter(recipeIDs, function (recipeID)
             return C_TradeSkillUI.IsRecipeInSkillLine(recipeID, professionID)
         end)
-    elseif filter == self.Filter.Tracked then
+    elseif filter == self.Filter.Queue then
         return Util:TblFilter(recipeIDs, Recipes.IsTracked, false, Recipes)
     elseif filter == self.Filter.Restock then
         return Util:TblFilter(recipeIDs, Restock.IsTracked, false, Restock)
@@ -383,14 +460,14 @@ end
 function Self:GetFilterRecipeOperations(filter, recipe, method)
     ---@type table<number, Operation | false>
     local operations
-    local Service = Util:Select(filter, self.Filter.Tracked, Recipes, self.Filter.Restock, Restock)
+    local Service = Util:Select(filter, self.Filter.Queue, Recipes, self.Filter.Restock, Restock)
 
     if Service then
         local amounts = Service:GetTrackedAmounts(recipe) ---@cast amounts -?
 
         operations = Util:TblMap(amounts, Util.FnFalse) --[[@as table<number, Operation | false>]]
 
-        if filter == self.Filter.Tracked then
+        if filter == self.Filter.Queue then
             for quality,amount in pairs(amounts) do repeat
                 if amount == 0 then break end
                 operations[quality] = Recipes:GetTrackedAllocation(recipe, quality) or false
@@ -404,16 +481,18 @@ function Self:GetFilterRecipeOperations(filter, recipe, method)
     local key, time = cache:Key(method, recipe), Prices:GetRecipeScanTime(recipe)
 
     if not cache:Has(key) or cache:Get(key)[1] ~= time then
-        local includeNonTradable = filter == self.Filter.Tracked
+        local includeNonTradable = filter == self.Filter.Queue
         cache:Set(key, { time, Optimization:GetRecipeAllocations(recipe, method, includeNonTradable) })
     end
 
     local optimized = cache:Get(key)[2]
     if not operations then return optimized end
 
+    local minQuality = optimized and Util:TblMinKey(optimized)
+
     for quality,op in pairs(operations) do repeat
         if op then break end
-        operations[quality] = optimized and optimized[quality]
+        operations[quality] = optimized and optimized[max(quality, minQuality)]
     until true end
 
     return operations
@@ -427,14 +506,16 @@ end
 ---@return number? amountShown
 ---@return number? amountTotal
 function Self:GetFilterRecipeAmount(filter, recipe, quality, value)
-    if filter == self.Filter.Tracked then
+    if filter == self.Filter.Queue then
         return Recipes:GetTrackedAmount(recipe, quality) --[[@as number]]
     elseif filter == self.Filter.Restock then
         local total = Restock:GetTrackedAmount(recipe, quality)
 
-        if value < Restock:GetTrackedMinProfit(recipe, quality or 1) then return 0, total end
+        if Prices:IsSourceInstalled() and value < Restock:GetTrackedMinProfit(recipe, quality or 1) then
+            return 0, total
+        end
 
-        return Restock:GetTrackedMissing(recipe), total
+        return Restock:GetTrackedMissing(recipe, quality), total
     end
 end
 
@@ -444,12 +525,13 @@ end
 
 -- Craft/Restock
 
+---@return RecipeTreeNode?
 function Self:GetNextTrackedCraftable()
-    if self.filter ~= self.Filter.Tracked then return end
+    if self.filter ~= self.Filter.Queue then return end
     if self.dataProvider:IsEmpty() then return end
 
-    return select(2, self.dataProvider:FindByPredicate(function (node)
-        local data = node:GetData() --[[@as RecipeTreeNodeData]]
+    return select(2, self.dataProvider:FindByPredicate(function (node) ---@cast node RecipeTreeNode
+        local data = node:GetData()
         if not data.amount or not data.operation then return false end
         return data.amount > 0 and data.operation:HasAllReagents()
     end, false))
@@ -465,48 +547,54 @@ function Self:GetNextTrackedCRestock()
 end
 
 function Self:CraftRestockButtonOnClick()
-    if self.filter == self.Filter.Tracked then
+    if self.filter == self.Filter.Queue then
         local node = self:GetNextTrackedCraftable()
         if not node then return end
 
-        local data = node:GetData() --[[@as RecipeTreeNodeData]]
-        local operation = data.operation
-        local amount = min(data.amount, operation:GetMaxCraftAmount())
+        local data = node:GetData()
+        local operation, amount = data.operation, data.amount
+        if not operation or not amount then return end
+
+         amount = min(amount, operation:GetMaxCraftAmount())
 
         self:CraftOperation(operation, amount)
     elseif self.filter == self.Filter.Restock then
-        for _,node in self.dataProvider:EnumerateEntireRange() do repeat
-            local data = node:GetData() --[[@as RecipeTreeNodeData]]
+        for _,node in self.dataProvider:EnumerateEntireRange() do repeat ---@cast node RecipeTreeNode
+            local data = node:GetData()
             local operation, quality, amount = data.operation, data.quality, data.amount
-            local recipe = operation.recipe
+            if not operation or not amount then break end
 
             if amount <= 0 then break end
 
-            Recipes:SetTracked(recipe)
-            Recipes:SetTrackedPerQuality(recipe, quality ~= nil)
-            Recipes:SetTrackedAmount(recipe, amount, quality)
-            Recipes:SetTrackedAllocation(recipe, operation, quality)
+            Recipes:SetTracked(operation.recipe)
+            Recipes:SetTrackedPerQuality(operation.recipe, quality ~= nil)
+            Recipes:SetTrackedAmount(operation.recipe, amount, quality)
+            Recipes:SetTrackedAllocation(operation.recipe, operation, quality)
         until true end
     end
 end
 
 -- Prev/Next filter
 
-local filterList = { Self.Filter.Sorted, Self.Filter.Tracked, Self.Filter.Restock }
-
 function Self:PrevFilterButtonOnClick()
-    local curr = Util:TblIndexOf(filterList, self.filter) or 0
-    self:SetFilterSelected(filterList[curr - 1])
+    local curr = Util:TblIndexOf(self.filterList, self.filter) or 0
+    self:SetFilterSelected(self.filterList[curr - 1])
 end
 
 function Self:NextFilterButtonOnClick()
-    local curr = Util:TblIndexOf(filterList, self.filter) or 0
-    self:SetFilterSelected(filterList[curr + 1])
+    local curr = Util:TblIndexOf(self.filterList, self.filter) or 0
+    self:SetFilterSelected(self.filterList[curr + 1])
 end
 
 -- Insert/Update buttons
 
 function Self:InsertFilterButtons()
+    if Prices:IsSourceInstalled() then
+        self.filterList = { self.Filter.Scan, self.Filter.Queue, self.Filter.Restock }
+    else
+        self.filterList = { self.Filter.Queue, self.Filter.Restock }
+    end
+
     self.craftRestockBtn = GUI:InsertButton(
         "", self.frame, nil, Util:FnBind(self.CraftRestockButtonOnClick, self),
         "TOP", self.frame.RecipeList, "BOTTOM", 0, -3.5
@@ -523,23 +611,20 @@ function Self:InsertFilterButtons()
         "TOPRIGHT", self.frame.RecipeList, "BOTTOMRIGHT", -1, -3.5
     )
 
-    self.prevFilterBtn.tooltipText = "Previous filter page"
-    self.nextFilterBtn.tooltipText = "Next filter page"
-
     self.craftRestockBtn:Hide()
     self.prevFilterBtn:Hide()
     self.nextFilterBtn:Hide()
 end
 
 function Self:UpdateFilterButtons()
-    local shown = Util:OneOf(self.filter, self.Filter.Tracked, self.Filter.Restock)
+    local shown = Util:OneOf(self.filter, self.Filter.Queue, self.Filter.Restock)
     local enabled = shown and not self.filterJob
     local text = ""
 
     if self.filter == self.Filter.Restock then
         enabled = enabled and self:GetNextTrackedCRestock() ~= nil
         text = "Restock"
-    elseif self.filter == self.Filter.Tracked then
+    elseif self.filter == self.Filter.Queue then
         enabled = enabled and self:GetNextTrackedCraftable() ~= nil
         text = "Create Next"
     end
@@ -549,12 +634,16 @@ function Self:UpdateFilterButtons()
     self.craftRestockBtn:SetTextToFit(text)
 
     local shown = self.filter ~= nil
-    local curr = Util:TblIndexOf(filterList, self.filter) or 0
+    local curr = Util:TblIndexOf(self.filterList, self.filter) or 0
+    local prev, next = self.filterList[curr - 1], self.filterList[curr + 1]
 
     self.prevFilterBtn:SetShown(shown)
     self.nextFilterBtn:SetShown(shown)
     self.prevFilterBtn:SetText(curr > 1 and "<" or "X")
-    self.nextFilterBtn:SetEnabled(curr < #filterList)
+    self.nextFilterBtn:SetEnabled(curr < #self.filterList)
+
+    self.prevFilterBtn.tooltipText = Util:TblIndexOf(self.Filter, prev) --[[@as string?]] or "Close filter"
+    self.nextFilterBtn.tooltipText = Util:TblIndexOf(self.Filter, next) --[[@as string?]]
 end
 
 ---------------------------------------
@@ -596,12 +685,12 @@ end
 --              Events
 ---------------------------------------
 
----@param node TreeNodeMixin
+---@param node RecipeTreeNode
 ---@param selected boolean
 function Self:OnSelectionChanged(node, selected)
     if not selected or not self.frame:IsVisible() then return end
 
-    local form, data = GUI.RecipeForm.CraftingForm, node:GetData() --[[@as RecipeTreeNodeData]]
+    local form, data = GUI.RecipeForm.CraftingForm, node:GetData()
 
     if data.operation then
         form:SetOperation(data.operation)
@@ -651,7 +740,7 @@ function Self:OnRefresh()
 end
 
 ---@param frame ProfessionsRecipeListRecipeFrame
----@param node TreeNodeMixin
+---@param node RecipeTreeNode
 function Self:OnRecipeListRecipeInitialized(frame, node)
     self:InitRecipeListRecipe(frame, node)
 end
@@ -661,7 +750,7 @@ function Self:OnProfessionStatsChanged()
 end
 
 function Self:OnTrackedRecipeUpdated()
-    if self.filter ~= self.Filter.Tracked or not self.frame:IsVisible() then return end
+    if self.filter ~= self.Filter.Queue or not self.frame:IsVisible() then return end
     self:UpdateRecipeList(true)
 end
 
@@ -671,7 +760,7 @@ function Self:OnTrackedRestockUpdated()
 end
 
 function Self:OnOwnedItemsUpdated()
-    if not Util:OneOf(self.filter, self.Filter.Restock, self.Filter.Tracked) or not self.frame:IsVisible() then return end
+    if not Util:OneOf(self.filter, self.Filter.Restock, self.Filter.Queue) or not self.frame:IsVisible() then return end
     self:UpdateRecipeList(true)
 end
 
@@ -685,8 +774,6 @@ function Self:OnAddonLoaded(addonName)
     hooksecurefunc(self.frame, "ValidateControls", Util:FnBind(self.ValidateControls, self))
 
     GUI:RegisterCallback(GUI.Event.Refresh, self.OnRefresh, self)
-
-    if not Prices:IsSourceInstalled() then return end
 
     self:CreateRecipeListProgressBar()
     self:ModifyRecipeListFilter()
@@ -728,4 +815,7 @@ Addon:RegisterCallback(Addon.Event.AddonLoaded, Self.OnAddonLoaded, Self)
 --              Types
 ---------------------------------------
 
----@alias RecipeTreeNodeData { recipeInfo: TradeSkillRecipeInfo, operation: Operation, value: number, method: Optimization.Method, quality: number, amount?: number, total?: number }
+---@class RecipeTreeNode: TreeNodeMixin
+---@field GetData fun(self: self): RecipeTreeNodeData
+
+---@alias RecipeTreeNodeData { recipeInfo: TradeSkillRecipeInfo, operation?: Operation, value?: number, method?: Optimization.Method, quality?: number, amount?: number, total?: number }